move func to dom
