+++ extensions/amp-ad/0.1/test/test-amp-ad-3p-impl.js
@@ -17,7 +17,7 @@
 import {AmpAd3PImpl} from '../amp-ad-3p-impl';
 import {createAdPromise} from '../../../../testing/ad-iframe';
 import {createIframePromise} from '../../../../testing/iframe';
-import {createElementWithAttributes} from '../../../../testing/test-helper';
+import {createElementWithAttributes} from '../../../../src/dom';
 import '../../../amp-sticky-ad/0.1/amp-sticky-ad';
 import * as sinon from 'sinon';
 import * as lolex from 'lolex';
@@ -33,7 +33,8 @@ function tests(name) {
 
   function getAdInAdContainer() {
     return createIframePromise().then(iframe => {
-      const adContainer = iframe.doc.createElement('amp-sticky-ad');
+      const adContainer = createElementWithAttributes(iframe.doc,
+          'amp-sticky-ad', {layout: 'nodisplay'});
       const ampAd = createElementWithAttributes(iframe.doc, 'amp-ad', {
         width: 300,
         height: 50,
